% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{dados_candidates}
\alias{dados_candidates}
\title{Dados de candidates}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 223 rows and 23 columns.
}
\usage{
dados_candidates
}
\description{
Tabela com alguns dados de candidaturas listados por AzMina
}
\keyword{datasets}
